// src/apiV2/a7-service/model/promptResponse.ts
/**
 * Generated by orval v6.6.2 🍺
 * Do not edit manually.
 * WanmaX Backend API
 * API документация для WanmaX Backend - Автоматизированный конвейер обработки фотографий
 * OpenAPI spec version: 1.0.0
 */
import type { PromptResponseHistoryItem } from './promptResponseHistoryItem';

export interface PromptResponse {
  /** Уникальный идентификатор промпта */
  id?: string;
  /** Название промпта */
  title?: string;
  /** Текст промпта */
  body?: string;
  /** Исходный промпт */
  default?: string;
  /** Версии промпта */
  history?: PromptResponseHistoryItem[];
  /** Дата создания */
  createdAt?: string;
  /** Дата последнего обновления */
  updatedAt?: string;
}